`Refactor pages to use server-side Supabase client` Connected the dashboard,pickup and reports.
