+++ src/components/Header.jsx
@@ -1,13 +1,11 @@
 import Link from 'next/link';
+import Brand from './Brand';
 
 export default function Header(){
   return (
     <header className="navbar">
       <div className="container nav-inner center" style={{justifyContent:'space-between'}}>
-        <div className="brand">
-          <span style={{display:'inline-flex',width:28,height:28,alignItems:'center',justifyContent:'center',borderRadius:8,background:'#15803d',color:'#fff',fontWeight:700}}>🌿</span>
-          <span>MyCommunityPortal</span>
-        </div>
+        <Brand />
         <nav className="nav-links">
           <a href="#features">Features</a>
           <a href="#how">How It Works</a>
