+++ src/lib/supabase/index.js
@@ -1,0 +1,2 @@
+export { supabase } from './supabase';
+
